Organized files, added tests for Cpu
