+++ package.json
@@ -10,11 +10,11 @@
     "bundle2": "watchify src/webpage/pages.js -o public/limited.js --debug",
     "eslint": "eslint src/ --fix",
     "node": "node server.js",
-    "sass": "sass --watch src/css/style.scss temp.css --no-source-map",
-    "sass_prod": "sass src/css/style.scss temp.css --no-source-map",
+    "sass": "sass --watch src/_css/style.scss temp.css --no-source-map",
+    "sass_prod": "sass src/_css/style.scss temp.css --no-source-map",
     "start": "npm-run-all --parallel bundle bundle2 node sass autoprefix",
-    "stylelint": "stylelint \"src/css/*.scss\" --fix",
-    "test": "eslint src/ --fix && stylelint \"src/css/*.scss\" --fix && mocha",
+    "stylelint": "stylelint \"src/_css/*.scss\" --fix",
+    "test": "eslint src/ --fix && stylelint \"src/_css/*.scss\" --fix && mocha 'test/**/*.test.js'",
     "uglify": "browserify -t uglifyify src/Main.js > public/bundle.js",
     "uglify_limited": "browserify -t uglifyify src/webpage/pages.js > public/limited.js"
   },
