+++ src/cpu/CpuVariants.js
@@ -1,105 +1,26 @@
 'use strict';
 
-const { Board } = require('./Board.js');
+const { Board } = require('../Board.js');
+const { Cpu } = require('./Cpu.js');
 
-class Cpu {
-	constructor(settings) {
-		if(this.constructor === Cpu) {
-			throw new Error('Abstract class cannot be instatiated.');
+class CpuVariants {
+	constructor() {
+		throw new Error('CpuVariants cannot be instantiated.');
+	}
+
+	static fromString(ai, settings) {
+		switch(ai.toLowerCase()) {
+			case 'random':
+				return new RandomCpu(settings);
+			case 'flat':
+				return new FlatCpu(settings);
+			case 'tall':
+				return new TallCpu(settings);
+			case 'chain':
+				return new ChainCpu(settings);
+			default:
+				return new TestCpu(settings);
 		}
-		this.settings = settings;
-	}
-
-	/**
-	 * Returns the optimal move according to the AI.
-	 */
-	/* eslint-disable-next-line no-unused-vars*/
-	getMove(boardState, currentDrop) {
-		throw new Error('getMove(boardState, currentDrop) must be implemented by the subclass.');
-	}
-
-	/**
-	 * Returns the average height of all the columns.
-	 */
-	getAverageHeight(boardState) {
-		return boardState.reduce((sum, col) => sum += col.length, 0) / this.settings.cols;
-	}
-
-	/**
-	 * Returns the best column placement with either 0 or 2 rotations that makes a chain longer than minChain.
-	 * If none exist, returns -1.
-	 */
-	checkForSimpleChains(boardState, currentDrop, minChain) {
-		let runningMaxChain = minChain;
-		let col = -1;
-		for(let i = 0; i < this.settings.cols * 2; i++) {
-			const currCol = Math.floor(i / 2);
-			const board = new Board(this.settings, boardState);
-			if(i % 2 === 0) {
-				board.boardState[currCol].push(currentDrop.colours[0]);
-				board.boardState[currCol].push(currentDrop.colours[1]);
-			}
-			else {
-				board.boardState[currCol].push(currentDrop.colours[1]);
-				board.boardState[currCol].push(currentDrop.colours[0]);
-			}
-
-			const chains = board.resolveChains();
-			if(chains.length > runningMaxChain) {
-				runningMaxChain = chains.length;
-				col = currCol;
-			}
-		}
-		return col;
-	}
-
-	/**
-	 * Returns the move that results in the best chain longer than minChain.
-	 * If none exist, returns { col: -1, rotations: -1 };
-	 */
-	checkForAllChains(boardState, currentDrop, minChain) {
-		let runningMaxChain = minChain;
-		let col = -1;
-		let rotations = -1;
-		for(let i = 0; i < this.settings.cols * 4; i++) {
-			const currCol = i % this.settings.cols;
-			const board = new Board(this.settings, boardState);
-			let tempRotations;
-			if(i < this.settings.cols) {
-				board.boardState[currCol].push(currentDrop.colours[1]);
-				board.boardState[currCol].push(currentDrop.colours[0]);
-				tempRotations = 2;
-			}
-			else if(i < this.settings.cols * 2) {
-				if(currCol === 0) {
-					continue;
-				}
-				board.boardState[currCol - 1].push(currentDrop.colours[0]);
-				board.boardState[currCol].push(currentDrop.colours[1]);
-				tempRotations = -1;
-			}
-			else if(i < this.settings.cols * 3) {
-				if(currCol === this.settings.cols - 1) {
-					continue;
-				}
-				board.boardState[currCol].push(currentDrop.colours[0]);
-				board.boardState[currCol + 1].push(currentDrop.colours[1]);
-				tempRotations = 1;
-			}
-			else {
-				board.boardState[currCol].push(currentDrop.colours[0]);
-				board.boardState[currCol].push(currentDrop.colours[1]);
-				tempRotations = 0;
-			}
-
-			const chains = board.resolveChains();
-			if(chains.length > runningMaxChain) {
-				runningMaxChain = chains.length;
-				col = currCol;
-				rotations = tempRotations;
-			}
-		}
-		return { col, rotations };
 	}
 
 	static getAllCpuNames() {
@@ -111,23 +32,7 @@ class Cpu {
 			'Test'
 		];
 	}
-
-	static fromString(ai, settings) {
-		switch(ai) {
-			case 'Random':
-				return new RandomCpu(settings);
-			case 'Flat':
-				return new FlatCpu(settings);
-			case 'Tall':
-				return new TallCpu(settings);
-			case 'Chain':
-				return new ChainCpu(settings);
-			default:
-				return new TestCpu(settings);
-		}
-	}
 }
-
 
 /**
  * RandomCpu: Completely random moves.
@@ -155,8 +60,9 @@ class FlatCpu extends Cpu {
 
 	getMove(boardState, currentDrop) {
 		let col = 0;
-		const rotations = 0;
+		let rotations = 0;
 		let minHeight = -1;
+
 		for(let i = 0; i < this.settings.cols - 1; i++) {
 			if(boardState[i].length < minHeight) {
 				minHeight = boardState[i].length;
@@ -166,8 +72,9 @@ class FlatCpu extends Cpu {
 
 		const result = super.checkForSimpleChains(boardState, currentDrop, 1);
 
-		if(result !== -1) {
-			col = result;
+		if(result.col !== -1) {
+			col = result.col;
+			rotations = result.rotations;
 		}
 
 		return { col, rotations };
@@ -229,7 +136,7 @@ class ChainCpu extends Cpu {
 
 	getMove(boardState, currentDrop) {
 		let col = Math.floor(Math.random() * this.settings.cols);
-		const rotations = 0;
+		let rotations = 0;
 
 		// Deter against random placements in column 2 (when 0-indexed)
 		while(col === 2) {
@@ -238,8 +145,9 @@ class ChainCpu extends Cpu {
 
 		const result = super.checkForSimpleChains(boardState, currentDrop, 1);
 
-		if(result !== -1) {
-			col = result;
+		if(result.col !== -1) {
+			col = result.col;
+			rotations = result.rotations;
 		}
 
 		return { col, rotations };
@@ -260,7 +168,7 @@ class TestCpu extends Cpu {
 						(averageHeight > this.settings.rows / 2) ? 2 :		// eslint-disable-line indent
 						(averageHeight > this.settings.rows / 2) ? 3 : 4;	// eslint-disable-line indent
 
-		let { col, rotations} = super.checkForAllChains(boardState, currentDrop, minChain);
+		let { col, rotations } = super.checkForAllChains(boardState, currentDrop, minChain);
 
 		// Unable to find an appropriate chain
 		if(col === -1) {
@@ -367,11 +275,4 @@ class TestCpu extends Cpu {
 	}
 }
 
-module.exports = {
-	Cpu,
-	RandomCpu,
-	FlatCpu,
-	TallCpu,
-	ChainCpu,
-	TestCpu
-};
+module.exports = { CpuVariants };
