+++ src/utils/Utils.js
@@ -1,0 +1,115 @@
+'use strict';
+
+/**
+ * Returns a random puyo colour, given the size of the colour pool.
+ */
+function getRandomColour (numColours) {
+	return Math.floor(Math.random() * numColours) + 1;
+}
+
+/**
+ * Returns the location(s) of the schezo puyo(s).
+ *
+ * Currently only works for I-shaped Drops (Tsu).
+ */
+function getOtherPuyo (drop) {
+	let x = drop.arle.x + Math.cos(drop.standardAngle + Math.PI / 2);
+	let y = drop.arle.y + Math.sin(drop.standardAngle + Math.PI / 2);
+
+	// Perform integer rounding
+	if(Math.abs(x - Math.round(x)) < 0.001) {
+		x = Math.round(x);
+	}
+	if(Math.abs(y - Math.round(y)) < 0.001) {
+		y = Math.round(y);
+	}
+	return { x, y };
+}
+
+/**
+ * Gets the frames needed for the animation (accounts for falling time).
+ */
+function getDropFrames(poppingLocs, boardState, settings) {
+	return poppingLocs.some(loc => {
+		return boardState[loc.col][loc.row + 1] !== undefined && !poppingLocs.some(loc2 => loc2.col === loc.col && loc2.row === loc.row + 1);
+	}) ? settings.dropFrames : 0;
+}
+
+/**
+ * Finds the score of the given chain. Currently only for Tsu rule.
+ */
+function calculateScore (puyos, chain_length) {
+	// These arrays are 1-indexed.
+	const CHAIN_POWER = [null, 0, 8, 16, 32, 64, 96, 128, 160, 192, 224, 256, 288, 320, 352, 384, 416, 448, 480, 512, 544, 576, 608, 640, 672];
+	const COLOUR_BONUS = [null, 0, 3, 6, 12, 24, 48];
+	const GROUP_BONUS = [null, null, null, null, 0, 2, 3, 4, 5, 6, 7, 10, 10, 10, 10];
+
+	// Number of puyos cleared in the chain
+	const puyos_cleared = puyos.length;
+
+	// Find the different colours
+	const containedColours = {};
+
+	puyos.forEach(puyo => {
+		if(containedColours[puyo.colour] === undefined) {
+			containedColours[puyo.colour] = 1;
+		}
+		else {
+			containedColours[puyo.colour]++;
+		}
+	});
+
+	// Chain power based on length of chain
+	const chain_power = CHAIN_POWER[chain_length];
+
+	// Colour bonus based on number of colours used
+	const colour_bonus = COLOUR_BONUS[Object.keys(containedColours).length];
+
+	// Group bonus based on number of puyos in each group
+	const group_bonus = Object.keys(containedColours).reduce((bonus, colour) => bonus += GROUP_BONUS[containedColours[colour]], 0);
+
+	return (10 * puyos_cleared) * (chain_power + colour_bonus + group_bonus);
+}
+
+/**
+ * Finds the amount of nuisance generated from a particular increase in score.
+ */
+function calculateNuisance(chain_score, targetPoints, leftoverNuisance) {
+	const nuisancePoints = chain_score / targetPoints + leftoverNuisance;
+	const nuisanceSent = Math.floor(nuisancePoints);
+
+	return { nuisanceSent, leftoverNuisance: nuisancePoints - nuisanceSent };
+}
+
+/**
+ * Deep copies an object where all values are primitype types.
+ * Call this function recursively to deep copy more nested objects.
+ */
+function objectCopy(obj) {
+	return JSON.parse(JSON.stringify(obj));
+}
+
+/**
+ * Clamps a number between a minimum and maximum number.
+ */
+function clampBetween(value, min, max) {
+	if(value < min) {
+		return min;
+	}
+	else if(value > max) {
+		return max;
+	}
+	return value;
+}
+
+const Utils = {
+	getRandomColour,
+	getOtherPuyo,
+	getDropFrames,
+	calculateScore,
+	calculateNuisance,
+	objectCopy,
+	clampBetween
+};
+
+module.exports = { Utils };
