+++ test/Board.test.js
@@ -3,18 +3,15 @@
 /*eslint-env mocha */
 
 const deepEqualInAnyOrder = require('deep-equal-in-any-order');
-const chaiExclude = require('chai-exclude');
-const { expect } = require('chai').use(deepEqualInAnyOrder).use(chaiExclude);
+const { expect } = require('chai').use(deepEqualInAnyOrder);
+
 const { Board } = require('../src/Board.js');
-const { Drop } = require('../src/Drop.js');
-const { Utils, Settings } = require('../src/Utils.js');
+const { Settings } = require('../src/utils/Settings.js');
 
 const red = 1;
 const grn = 2;
 const blu = 3;
 const nui = 0;
-
-const settings = new Settings();
 
 describe('Board.js', function() {
 	describe('getConnections', function() {
@@ -355,179 +352,6 @@ describe('Board.js', function() {
 
 			expect(nuisanceDropped).to.equal(30);
 			expect(total_nuisance).to.equal(30);
-		});
-	});
-});
-
-describe('Utils.js', function() {
-	describe('Settings', function() {
-		it('should be able to transform into a string and back', function() {
-			const settings1 = new Settings();
-			const settings2 = Settings.fromString(settings1.toString());
-
-			expect(settings1).excluding(['timer', 'seed']).to.deep.equal(settings2);
-		});
-
-		it('checkMarginTime - should not change target points before margin time', function() {
-			const settings2 = new Settings();
-			const startTime = settings2.timer;
-			const targetPoints = settings2.targetPoints;
-
-			settings2.checkMarginTime(startTime + 32000);
-
-			expect(settings2.targetPoints).to.equal(targetPoints);
-			expect(settings2.timer).to.equal(startTime);
-		});
-
-		it('checkMarginTime - should update target points correctly jumping into margin time', function() {
-			const settings2 = new Settings();
-			const startTime = settings2.timer;
-			const targetPoints = settings2.targetPoints;
-
-			// Start margin time and incrment twice
-			settings2.checkMarginTime(startTime + (96001 + 32000));
-			const expectedResult = Math.floor(Math.floor(Math.floor(targetPoints * 0.75) / 2) / 2);
-
-			expect(settings2.targetPoints).to.equal(expectedResult);
-			expect(settings2.timer).to.equal(startTime + (96000 + 32000));
-		});
-
-		it('checkMarginTime - should update target points correctly incrementing margin time', function() {
-			const settings2 = new Settings();
-			const startTime = settings2.timer;
-			const targetPoints = settings2.targetPoints;
-
-			// Start margin time
-			const midTime1 = startTime + 96001;
-			settings2.checkMarginTime(midTime1);
-			const expectedResultMid = Math.floor(targetPoints * 0.75);
-
-			expect(settings2.targetPoints).to.equal(expectedResultMid);
-			expect(settings2.timer).to.equal(midTime1 - 1);
-
-			// Margin time has started, but not yet re-incremented
-			const midTime2 = midTime1 + 9000;
-			settings2.checkMarginTime(midTime2);
-
-			expect(settings2.targetPoints).to.equal(expectedResultMid);
-			expect(settings2.timer).to.equal(midTime1 - 1);
-
-			// Margin time has started and incremented once
-			const finalTime = midTime2 + 8001;
-			settings2.checkMarginTime(finalTime);
-			const expectedResultFinal = Math.floor(expectedResultMid / 2);
-
-			expect(settings2.targetPoints).to.equal(expectedResultFinal);
-			expect(settings2.timer).to.equal(midTime1 + 16000 - 1);
-		});
-	});
-
-	describe('getOtherPuyo', function() {
-		it('should retrieve the correct location of a Tsu schezo', function() {
-			// Set the arle at (1, 1) with standardAngle 0 deg
-			const drop1 = new Drop('I', [red, blu], settings, { x: 1, y: 1}, undefined, 0);
-			const schezo1 = Utils.getOtherPuyo(drop1);
-
-			expect(schezo1).to.deep.equal({ x: 1, y: 2});
-
-			// Set the arle at (1, 1) with standardAngle 120 deg
-			const drop2 = new Drop('I', [red, blu], settings, { x: 1, y: 1}, undefined, 120);
-			const schezo2 = Utils.getOtherPuyo(drop2);
-
-			expect(schezo2.x).to.be.within(0.419, 0.42);
-			expect(schezo2.y).to.be.within(1.814, 1.815);
-		});
-	});
-
-	describe('getDropFrames', function() {
-		it('should return 0 when no puyos fall', function() {
-			const boardState = [
-				[red, red, blu, nui],
-				[red, blu, blu],
-				[grn, blu],
-				[blu],
-				[],
-				[]
-			];
-			const poppingLocs = [
-				{ col: 0, row: 2 },
-				{ col: 0, row: 3 },		// popped nuisance
-				{ col: 1, row: 1 },
-				{ col: 1, row: 2 },
-				{ col: 2, row: 1 }
-			];
-			const dropFrames = Utils.getDropFrames(poppingLocs, boardState, settings);
-			expect(dropFrames).to.equal(0);
-		});
-
-		it('should return the number of drop frames when puyos fall', function() {
-			const boardState = [
-				[red, red],
-				[red],
-				[red, grn],
-				[],
-				[],
-				[]
-			];
-			const poppingLocs = [
-				{ col: 0, row: 0 },
-				{ col: 0, row: 1 },
-				{ col: 1, row: 0 },
-				{ col: 2, row: 0 }
-			];
-			const dropFrames = Utils.getDropFrames(poppingLocs, boardState, settings);
-			expect(dropFrames).to.equal(settings.dropFrames);
-		});
-
-		it('should return ${dropFrames} when nuisance fall', function() {
-			const boardState = [
-				[red, red, nui, nui],
-				[red],
-				[red, grn],
-				[],
-				[],
-				[]
-			];
-			const poppingLocs = [
-				{ col: 0, row: 2 },
-				{ col: 0, row: 3 },		//popped nuisance
-				{ col: 1, row: 1 },
-				{ col: 1, row: 2 },
-				{ col: 2, row: 1 }
-			];
-			const dropFrames = Utils.getDropFrames(poppingLocs, boardState, settings);
-			expect(dropFrames).to.equal(0);
-		});
-	});
-
-	describe('calculateScore', function() {
-		it('should calculate the correct score for a group of 4', function() {
-			const puyos = [
-				{ colour: grn },
-				{ colour: grn },
-				{ colour: grn },
-				{ colour: grn }
-			];
-			// 4 puyos together as the 2nd chain
-			const score = Utils.calculateScore(puyos, 2);
-			expect(score).to.equal(4 * 80);
-		});
-
-		it('should calculate the correct score for a power chain', function() {
-			const puyos = [
-				{ colour: grn },
-				{ colour: grn },
-				{ colour: grn },
-				{ colour: grn },
-				{ colour: blu },
-				{ colour: blu },
-				{ colour: blu },
-				{ colour: blu },
-				{ colour: blu}
-			];
-			// 9 puyos together as the 4th chain
-			const score = Utils.calculateScore(puyos, 4);
-			expect(score).to.equal(90 * 37);
 		});
 	});
 });
